requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asyncer==0.0.2
attrs==23.2.0
bidict==0.23.1
boto3==1.34.94
botocore==1.34.94
certifi==2024.2.2
chainlit==1.0.505
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
Deprecated==1.2.14
distlib==0.3.8
fastapi==0.110.3
fastapi-socketio==0.0.10
filelock==3.13.4
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.3.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
importlib-metadata==7.0.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-aws==0.1.2
langchain-community==0.0.34
langchain-core==0.1.45
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.49
Lazify==0.4.0
literalai==0.0.507
MarkupSafe==2.1.5
marshmallow==3.21.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opensearch-py==2.5.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.1
packaging==23.2
pillow==10.3.0
platformdirs==4.2.0
protobuf==4.25.3
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-multipart==0.0.9
python-socketio==5.11.2
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
threadpoolctl==3.4.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.0
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
uptrace==1.24.0
urllib3==2.2.1
uvicorn==0.25.0
virtualenv==20.25.2
watchfiles==0.20.0
websockets==12.0
whitenoise==6.6.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
